refactor: split runtime types
