+++ src/runtime/composables/useStrapiClient.ts
@@ -1,7 +1,7 @@
 import type { FetchError, FetchOptions } from 'ohmyfetch'
 import { stringify } from 'qs'
-import type { Strapi4Error } from '../../types/v4'
-import type { Strapi3Error } from '../../types/v3'
+import type { Strapi4Error } from '../types/v4'
+import type { Strapi3Error } from '../types/v3'
 import { useStrapiUrl } from './useStrapiUrl'
 import { useStrapiVersion } from './useStrapiVersion'
 import { useStrapiToken } from './useStrapiToken'
