+++ src/runtime/composables/useStrapiUser.ts
@@ -1,5 +1,5 @@
 import type { Ref } from 'vue'
-import type { StrapiUser } from '../../types'
+import type { StrapiUser } from '../types'
 import { useState } from '#app'
 
 export const useStrapiUser = (): Ref<StrapiUser> => useState<StrapiUser>('strapi_user')
